+++ CHANGELOG.md
@@ -9,6 +9,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 - Improve CLI errors for Atomic-Server #465
 - Fix default config directory, set it again to `~/.config/atomic`. This accidentally was `~` since v0.32.0.
 - Fix flaky query test #468
+- Don't subscribe to external resources #470
 
 ## [v0.32.2] - 2022-06-20